+++ server/responses/query_credit_types.go
@@ -5,6 +5,6 @@ import "github.com/google/uuid"
 type QueryCreditTypesResponse struct {
 	ID          uuid.UUID `json:"id"`
 	Name        string    `json:"name"`
-	Amount      float32   `json:"amount"`
+	Amount      int32     `json:"amount"`
 	Description string    `json:"description,omitempty"`
 }
